+++ 2024-Competition-Code/src/main/java/frc/robot/RobotContainer.java
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.commands.*;
-import frc.robot.subsystems.PhotonVision.PhotonVision;
+import frc.robot.subsystems.PhotonVision.photonvision;
 import frc.robot.subsystems.swerve.SwerveDrive;
 import frc.robot.subsystems.swerve.SwerveModuleIO;
 import frc.robot.subsystems.swerve.SwerveModuleIOSim;
@@ -17,8 +17,6 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.DataLogManager;
 import edu.wpi.first.wpilibj.DriverStation;
 import org.photonvision.PhotonCamera;
-
-import java.lang.annotation.Target;
 
 public class RobotContainer {
 
@@ -173,7 +171,7 @@ public class RobotContainer {
     allFour = new TestFourModules(swerve, actualXbox);
 
     PhotonCamera camera = new PhotonCamera("Microsoft_LifeCam_HD-3000");
-    PhotonVision photonVision = new PhotonVision(camera);
+    photonvision photonVision = new photonvision(camera);
     targetAprilTag = new TargetAprilTag(photonVision, swerve);
 
     teleopCommandChooser.addOption("AprilTagTargetingTEST", targetAprilTag);
