+++ 2024-Competition-Code/src/main/java/frc/robot/commands/TargetAprilTag.java
@@ -8,7 +8,7 @@ import frc.robot.subsystems.swerve.SwerveDrive;
 
 
 public class TargetAprilTag extends Command {
-    private PhotonVision photonVision;
+    private photonvision photonVision;
     private SwerveDrive swerveDrive;
     public XboxController controller;
 
@@ -19,7 +19,7 @@ public class TargetAprilTag extends Command {
     public boolean rotationAligned = false;
     public boolean centered = false;
 
-    public TargetAprilTag(PhotonVision photonVision, SwerveDrive swerveDrive) {
+    public TargetAprilTag(photonvision photonVision, SwerveDrive swerveDrive) {
         this.photonVision = photonVision;
         this.swerveDrive = swerveDrive;
