+++ 2024-Competition-Code/src/main/java/frc/robot/subsystems/photonvision/PhotonVision.java
@@ -14,7 +14,7 @@ import org.photonvision.targeting.PhotonPipelineResult;
 import org.photonvision.targeting.PhotonTrackedTarget;
 
 
-public class PhotonVision extends SubsystemBase {
+public class photonvision extends SubsystemBase {
     public PhotonCamera camera;
     public PhotonPipelineResult result;
     public PhotonTrackedTarget target;
@@ -25,7 +25,7 @@ public class PhotonVision extends SubsystemBase {
     public Constants.PhotonVisionConstants c;
 
 
-    public PhotonVision(PhotonCamera camera) {
+    public photonvision(PhotonCamera camera) {
         this.camera = camera;
         PortForwarder.add(5800, "photonvision", 5800);
         this.camera.setPipelineIndex(0);
